more policy perm tests, fold into policy tests module
