+++ c7n/query.py
@@ -155,7 +155,8 @@ class QueryResourceManager(ResourceManager):
         perms.append('%s:%s' % (m.service, _napi(m.enum_spec[0])))
         if getattr(m, 'detail_spec', None):
             perms.append("%s:%s" % (m.service, _napi(m.detail_spec[0])))
-        perms.extend(perms)
+        if getattr(cls, 'permissions', None):
+            perms.extend(cls.permissions)
         return perms
 
     def resources(self, query=None):
